Extract EventActionButtons to dedupe calendar/share UI (closes #33)
Pulls the duplicated calendar-add and send/share button logic out of
EventCard, AllDayCard, and EventModal into a single EventActionButtons
component. All state (calOpen, sendOpen, showCheck), refs, click-outside
effects, and handler functions (handleCopyText, handleSendInvite,
flashCheck) now live in one place. A `compact` prop (default false)
handles the AllDayCard's smaller icon/padding variant. EventModal retains
its own Escape-key handler since that's modal-scoped, not action-button
behavior.
