+++ frontend/src/components/AllDayStrip.jsx
@@ -1,8 +1,7 @@
-import { useState, useRef, useEffect } from 'react'
-import { CalendarDays, Send, Check } from 'lucide-react'
-import { googleCalendarUrl, downloadIcal, shareCalendarInvite, formatEventText } from '../lib/calendarUtils'
+import { useState } from 'react'
 import { sortedSources } from '../lib/sources'
 import EventModal from './EventModal'
+import EventActionButtons from './EventActionButtons'
 
 export default function AllDayStrip({ events }) {
   if (!events || events.length === 0) return null
@@ -26,50 +25,8 @@ export default function AllDayStrip({ events }) {
 
 function AllDayCard({ event }) {
   const [modalOpen, setModalOpen] = useState(false)
-  const [calOpen, setCalOpen] = useState(false)
-  const [sendOpen, setSendOpen] = useState(false)
-  const [showCheck, setShowCheck] = useState(false)
-  const calRef = useRef(null)
-  const sendRef = useRef(null)
   const sources = sortedSources(event.sources)
   const primary = sources[0]
-
-  useEffect(() => {
-    if (!calOpen) return
-    function handleOutside(e) {
-      if (calRef.current && !calRef.current.contains(e.target)) setCalOpen(false)
-    }
-    document.addEventListener('mousedown', handleOutside)
-    return () => document.removeEventListener('mousedown', handleOutside)
-  }, [calOpen])
-
-  useEffect(() => {
-    if (!sendOpen) return
-    function handleOutside(e) {
-      if (sendRef.current && !sendRef.current.contains(e.target)) setSendOpen(false)
-    }
-    document.addEventListener('mousedown', handleOutside)
-    return () => document.removeEventListener('mousedown', handleOutside)
-  }, [sendOpen])
-
-  function flashCheck() {
-    setShowCheck(true)
-    setTimeout(() => setShowCheck(false), 1500)
-  }
-
-  async function handleCopyText(e) {
-    e.stopPropagation()
-    await navigator.clipboard.writeText(formatEventText(event))
-    setSendOpen(false)
-    flashCheck()
-  }
-
-  async function handleSendInvite(e) {
-    e.stopPropagation()
-    const result = await shareCalendarInvite(event)
-    setSendOpen(false)
-    if (result?.downloaded) flashCheck()
-  }
 
   return (
     <>
@@ -92,56 +49,7 @@ function AllDayCard({ event }) {
             <h3 className="text-sm font-semibold text-gray-900 leading-snug line-clamp-2 flex-1 min-w-0">{event.title}</h3>
           )}
           <div className="flex items-center gap-0.5 flex-shrink-0 ml-1">
-            <div className="relative" ref={calRef}>
-              <button
-                className="text-gray-400 hover:text-gray-600 p-0.5 rounded cursor-pointer"
-                onClick={e => { e.stopPropagation(); setCalOpen(o => !o); setSendOpen(false) }}
-                title="Add to calendar"
-              >
-                <CalendarDays size={13} />
-              </button>
-              {calOpen && (
-                <div className="absolute right-0 top-6 z-10 bg-white border border-gray-200 rounded shadow-md text-sm min-w-max">
-                  <button
-                    className="block w-full text-left px-3 py-2 hover:bg-gray-50 whitespace-nowrap"
-                    onClick={e => { e.stopPropagation(); window.open(googleCalendarUrl(event), '_blank'); setCalOpen(false) }}
-                  >
-                    Google Calendar
-                  </button>
-                  <button
-                    className="block w-full text-left px-3 py-2 hover:bg-gray-50 whitespace-nowrap"
-                    onClick={e => { e.stopPropagation(); downloadIcal(event); setCalOpen(false) }}
-                  >
-                    Apple / Outlook (.ics)
-                  </button>
-                </div>
-              )}
-            </div>
-            <div className="relative" ref={sendRef}>
-              <button
-                className={`p-0.5 rounded cursor-pointer ${showCheck ? 'text-green-600' : 'text-gray-400 hover:text-gray-600'}`}
-                onClick={e => { e.stopPropagation(); setSendOpen(o => !o); setCalOpen(false) }}
-                title="Send / share event"
-              >
-                {showCheck ? <Check size={13} /> : <Send size={13} />}
-              </button>
-              {sendOpen && (
-                <div className="absolute right-0 top-6 z-10 bg-white border border-gray-200 rounded shadow-md text-sm min-w-max">
-                  <button
-                    className="block w-full text-left px-3 py-2 hover:bg-gray-50 whitespace-nowrap"
-                    onClick={handleCopyText}
-                  >
-                    Copy text
-                  </button>
-                  <button
-                    className="block w-full text-left px-3 py-2 hover:bg-gray-50 whitespace-nowrap"
-                    onClick={handleSendInvite}
-                  >
-                    Calendar invite
-                  </button>
-                </div>
-              )}
-            </div>
+            <EventActionButtons event={event} compact />
           </div>
         </div>
         {event.venue_name && (
