+++ frontend/src/components/EventModal.jsx
@@ -1,36 +1,13 @@
-import { useState, useRef, useEffect } from 'react'
+import { useEffect } from 'react'
 import { createPortal } from 'react-dom'
-import { CalendarDays, Send, Check, X } from 'lucide-react'
+import { X } from 'lucide-react'
 import { formatTimeRange } from '../lib/eventTime'
-import { googleCalendarUrl, downloadIcal, shareCalendarInvite, formatEventText } from '../lib/calendarUtils'
 import { sortedSources } from '../lib/sources'
+import EventActionButtons from './EventActionButtons'
 
 export default function EventModal({ event, onClose }) {
-  const [calOpen, setCalOpen] = useState(false)
-  const [sendOpen, setSendOpen] = useState(false)
-  const [showCheck, setShowCheck] = useState(false)
-  const calRef = useRef(null)
-  const sendRef = useRef(null)
   const sources = sortedSources(event.sources)
   const primaryUrl = sources[0]?.source_url
-
-  useEffect(() => {
-    if (!calOpen) return
-    function handleOutside(e) {
-      if (calRef.current && !calRef.current.contains(e.target)) setCalOpen(false)
-    }
-    document.addEventListener('mousedown', handleOutside)
-    return () => document.removeEventListener('mousedown', handleOutside)
-  }, [calOpen])
-
-  useEffect(() => {
-    if (!sendOpen) return
-    function handleOutside(e) {
-      if (sendRef.current && !sendRef.current.contains(e.target)) setSendOpen(false)
-    }
-    document.addEventListener('mousedown', handleOutside)
-    return () => document.removeEventListener('mousedown', handleOutside)
-  }, [sendOpen])
 
   useEffect(() => {
     function handleKey(e) {
@@ -39,25 +16,6 @@ export default function EventModal({ event, onClose }) {
     document.addEventListener('keydown', handleKey)
     return () => document.removeEventListener('keydown', handleKey)
   }, [onClose])
-
-  function flashCheck() {
-    setShowCheck(true)
-    setTimeout(() => setShowCheck(false), 1500)
-  }
-
-  async function handleCopyText(e) {
-    e.stopPropagation()
-    await navigator.clipboard.writeText(formatEventText(event))
-    setSendOpen(false)
-    flashCheck()
-  }
-
-  async function handleSendInvite(e) {
-    e.stopPropagation()
-    const result = await shareCalendarInvite(event)
-    setSendOpen(false)
-    if (result?.downloaded) flashCheck()
-  }
 
   return createPortal(
     <div
@@ -91,56 +49,7 @@ export default function EventModal({ event, onClose }) {
               <p className="text-xs text-gray-400 mt-0.5">{formatTimeRange(event.start_at, event.end_at)}</p>
             )}
             <div className="flex items-center gap-0.5 flex-shrink-0">
-              <div className="relative" ref={calRef}>
-                <button
-                  className="text-gray-400 hover:text-gray-600 p-1 rounded cursor-pointer"
-                  onClick={e => { e.stopPropagation(); setCalOpen(o => !o); setSendOpen(false) }}
-                  title="Add to calendar"
-                >
-                  <CalendarDays size={14} />
-                </button>
-                {calOpen && (
-                  <div className="absolute right-0 top-6 z-10 bg-white border border-gray-200 rounded shadow-md text-sm min-w-max">
-                    <button
-                      className="block w-full text-left px-3 py-2 hover:bg-gray-50 whitespace-nowrap"
-                      onClick={e => { e.stopPropagation(); window.open(googleCalendarUrl(event), '_blank'); setCalOpen(false) }}
-                    >
-                      Google Calendar
-                    </button>
-                    <button
-                      className="block w-full text-left px-3 py-2 hover:bg-gray-50 whitespace-nowrap"
-                      onClick={e => { e.stopPropagation(); downloadIcal(event); setCalOpen(false) }}
-                    >
-                      Apple / Outlook (.ics)
-                    </button>
-                  </div>
-                )}
-              </div>
-              <div className="relative" ref={sendRef}>
-                <button
-                  className={`p-1 rounded cursor-pointer ${showCheck ? 'text-green-600' : 'text-gray-400 hover:text-gray-600'}`}
-                  onClick={e => { e.stopPropagation(); setSendOpen(o => !o); setCalOpen(false) }}
-                  title="Send / share event"
-                >
-                  {showCheck ? <Check size={14} /> : <Send size={14} />}
-                </button>
-                {sendOpen && (
-                  <div className="absolute right-0 top-6 z-10 bg-white border border-gray-200 rounded shadow-md text-sm min-w-max">
-                    <button
-                      className="block w-full text-left px-3 py-2 hover:bg-gray-50 whitespace-nowrap"
-                      onClick={handleCopyText}
-                    >
-                      Copy text
-                    </button>
-                    <button
-                      className="block w-full text-left px-3 py-2 hover:bg-gray-50 whitespace-nowrap"
-                      onClick={handleSendInvite}
-                    >
-                      Calendar invite
-                    </button>
-                  </div>
-                )}
-              </div>
+              <EventActionButtons event={event} />
               <button
                 className="text-gray-400 hover:text-gray-600 p-1 rounded cursor-pointer ml-1"
                 onClick={onClose}
